+++ qgroundcontrol.qrc
@@ -59,7 +59,7 @@
         <file alias="LogReplaySettings.qml">src/UI/preferences/LogReplaySettings.qml</file>
         <file alias="MainRootWindow.qml">src/UI/MainRootWindow.qml</file>
         <file alias="MapSettings.qml">src/UI/preferences/MapSettings.qml</file>
-        <file alias="MavlinkConsolePage.qml">src/AnalyzeView/MavlinkConsolePage.qml</file>
+        <file alias="MAVLinkConsolePage.qml">src/AnalyzeView/MAVLinkConsolePage.qml</file>
         <file alias="MAVLinkInspectorPage.qml">src/AnalyzeView/MAVLinkInspectorPage.qml</file>
         <file alias="PX4LogTransferSettings.qml">src/UI/preferences/PX4LogTransferSettings.qml</file>
         <file alias="MissionSettingsEditor.qml">src/PlanView/MissionSettingsEditor.qml</file>
