+++ src/AnalyzeView/CMakeLists.txt
@@ -1,16 +1,28 @@
-find_package(Qt6 REQUIRED COMPONENTS Core Charts Gui Qml QmlIntegration Widgets)
+find_package(Qt6 REQUIRED COMPONENTS Core Charts Gui Qml QmlIntegration)
 
 qt_add_library(AnalyzeView STATIC
 	ExifParser.cc
 	ExifParser.h
 	GeoTagController.cc
 	GeoTagController.h
+	GeoTagWorker.cc
+	GeoTagWorker.h
 	LogDownloadController.cc
 	LogDownloadController.h
-	MavlinkConsoleController.cc
-	MavlinkConsoleController.h
+	LogEntry.cc
+	LogEntry.h
+	MAVLinkChartController.cc
+	MAVLinkChartController.h
+	MAVLinkConsoleController.cc
+	MAVLinkConsoleController.h
 	MAVLinkInspectorController.cc
 	MAVLinkInspectorController.h
+	MAVLinkMessage.cc
+	MAVLinkMessage.h
+	MAVLinkMessageField.cc
+	MAVLinkMessageField.h
+	MAVLinkSystem.cc
+	MAVLinkSystem.h
 	PX4LogParser.cc
 	PX4LogParser.h
 	ULogParser.cc
@@ -22,7 +34,6 @@ target_link_libraries(AnalyzeView
 		Qt6::Charts
 		Qt6::Gui
 		Qt6::Qml
-		Qt6::Widgets
 		FactSystem
 		QGC
 		QGCLocation
@@ -36,7 +47,7 @@ target_link_libraries(AnalyzeView
 		QmlControls
 )
 
-target_include_directories(AnalyzeView INTERFACE ${CMAKE_CURRENT_SOURCE_DIR})
+target_include_directories(AnalyzeView PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
 
 # qt_add_qml_module(AnalyzeView
 #     URI QGroundControl.AnalyzeView
@@ -46,14 +57,14 @@ target_include_directories(AnalyzeView INTERFACE ${CMAKE_CURRENT_SOURCE_DIR})
 # 		AnalyzeView.qml
 # 		GeoTagPage.qml
 # 		LogDownloadPage.qml
-# 		MavlinkConsolePage.qml
+# 		MAVLinkConsolePage.qml
 # 		MAVLinkInspectorPage.qml
 # 		VibrationPage.qml
 #     RESOURCES
 #     	FloatingWindow.svg
 #     	GeoTagIcon.svg
 #     	LogDownloadIcon.svg
-#     	MavlinkConsoleIcon.svg
+#     	MAVLinkConsoleIcon.svg
 #     	MAVLinkInspector.svg
 #     	VibrationPageIcon.png
 #     OUTPUT_TARGETS AnalyzeView_targets
