+++ src/AnalyzeView/ExifParser.cc
@@ -1,17 +1,10 @@
 #include "ExifParser.h"
+#include "QGCLoggingCategory.h"
+
 #include <QtCore/QtEndian>
 #include <QtCore/QDateTime>
 
-
-ExifParser::ExifParser()
-{
-
-}
-
-ExifParser::~ExifParser()
-{
-
-}
+QGC_LOGGING_CATEGORY(ExifParserLog, "qgc.analyzeview.exifparser")
 
 double ExifParser::readTime(QByteArray& buf)
 {
@@ -37,17 +30,17 @@ double ExifParser::readTime(QByteArray& buf)
 
     QStringList createDateList = createDate.split(' ');
     if (createDateList.count() < 2) {
-        qWarning() << "Could not decode creation time and date: " << createDateList;
+        qCWarning(ExifParserLog) << "Could not decode creation time and date: " << createDateList;
         return -1.0;
     }
     QStringList dateList = createDateList[0].split(':');
     if (dateList.count() < 3) {
-        qWarning() << "Could not decode creation date: " << dateList;
+        qCWarning(ExifParserLog) << "Could not decode creation date: " << dateList;
         return -1.0;
     }
     QStringList timeList = createDateList[1].split(':');
     if (timeList.count() < 3) {
-        qWarning() << "Could not decode creation time: " << timeList;
+        qCWarning(ExifParserLog) << "Could not decode creation time: " << timeList;
         return -1.0;
     }
     QDate date(dateList[0].toInt(), dateList[1].toInt(), dateList[2].toInt());
@@ -71,54 +64,6 @@ bool ExifParser::write(QByteArray& buf, GeoTagWorker::cameraFeedbackPacket& geot
     conversionPointer = reinterpret_cast<uint16_t *>(buf.mid(nextIfdOffsetInd, 2).data());
     uint16_t nextIfdOffset = *conversionPointer;
 
-    // Definition of useful unions and structs
-    union char2uint32_u {
-        char c[4];
-        uint32_t i;
-    };
-    union char2uint16_u {
-        char c[2];
-        uint16_t i;
-    };
-    // This struct describes a standart field used in exif files
-    struct field_s {
-        uint16_t tagID;  // Describes which information is added here, e.g. GPS Lat
-        uint16_t type;  // Describes the data type, e.g. string, uint8_t,...
-        uint32_t size;  // Describes the size
-        uint32_t content;  // Either contains the information, or the offset to the exif header where the information is stored (if 32 bits is not enough)
-    };
-    // This struct contains all the fields that we want to add to the image
-    struct fields_s {
-        field_s gpsVersion;
-        field_s gpsLatRef;
-        field_s gpsLat;
-        field_s gpsLonRef;
-        field_s gpsLon;
-        field_s gpsAltRef;
-        field_s gpsAlt;
-        field_s gpsMapDatum;
-        uint32_t finishedDataField;
-    };
-    // These are the additional information that can not be put into a single uin32_t
-    struct extended_s {
-        uint32_t gpsLat[6];
-        uint32_t gpsLon[6];
-        uint32_t gpsAlt[2];
-        char mapDatum[7];// = {0x57,0x47,0x53,0x2D,0x38,0x34,0x00};
-    };
-    // This struct contains all the information we want to add to the image
-    struct readable_s {
-        fields_s fields;
-        extended_s extendedData;
-    };
-
-    // This union is used because for writing the information we have to use a char array, but we still want the information to be available in a more descriptive way
-    union {
-        char c[0xa3];
-        readable_s readable;
-    } gpsData;
-
-
     char2uint32_u gpsIFDInd;
     gpsIFDInd.i = nextIfdOffset;
 
@@ -131,6 +76,8 @@ bool ExifParser::write(QByteArray& buf, GeoTagWorker::cameraFeedbackPacket& geot
 
     // filling values to gpsData
     uint32_t gpsDataExtInd = gpsIFDInd.i + 2 + sizeof(fields_s);
+
+    gpsData_u gpsData;
 
     // Filling up the fields with the corresponding values
     gpsData.readable.fields.gpsVersion.tagID = 0;
